Refactor size management and measurement data handling
- Updated SizeItem interface to include size analysis response.
- Modified AISuggestSizeType to accommodate new size analysis structure.
- Enhanced Dashboard and Measurements components to correctly handle and display measurement data.
- Improved layout structure in RootLayout for better context management.
- Added ListMeasure component for displaying measurement history with enhanced UI.
- Adjusted data fetching logic in AuthProvider to streamline user measurement data retrieval.
- Implemented error handling and loading states in AISuggestSizeProvider for image analysis.
- Ensured consistent data structure across components for better maintainability.

+++ Backend/Raidexi/Domain/Entities/SizeAnalysisResponse.cs
@@ -14,15 +14,15 @@ namespace Raidexi.Models
         public string Size_Us { get; set; }
         public int Size_Uk { get; set; }
         public int Size_Eu { get; set; }
-        public int Chest_Min_Cm { get; set; }
-        public int Chest_Max_Cm { get; set; }
-        public int Height_Min_Cm { get; set; }
-        public int Height_Max_Cm { get; set; }
-        public int? Weight_Min_Kg { get; set; }
-        public int? Weight_Max_Kg { get; set; }
-        public int Waist_Min_Cm { get; set; }
-        public int Waist_Max_Cm { get; set; }
-        public int Hip_Min_Cm { get; set; }
-        public int Hip_Max_Cm { get; set; }
+        public float  Chest_Min_Cm { get; set; }
+        public float Chest_Max_Cm { get; set; }
+        public float Height_Min_Cm { get; set; }
+        public float Height_Max_Cm { get; set; }
+        public float? Weight_Min_Kg { get; set; }
+        public float? Weight_Max_Kg { get; set; }
+        public float Waist_Min_Cm { get; set; }
+        public float Waist_Max_Cm { get; set; }
+        public float Hip_Min_Cm { get; set; }
+        public float Hip_Max_Cm { get; set; }
     }
 }
